Display unknown name as underscores
If given name or surname is blank, display it as a string of underscores so that it is clear which part of the name is unknown. However, do not allow the user to enter underscores - this is just for display. Prevent entry of any text recognized to mean "unknown name" (and require existing such text to be removed).